+++ src/search.mli
@@ -20,3 +20,4 @@ exception SearchFailure of string
 
 val search_document : string -> string -> string -> Doc.t list
 val get_database_name : string -> string
+val get_bibtex_from_doi : string -> string option