+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## v1.3.3
+* add ability to search for and copy BibTex from DOI
+* add field for notes to each entry
+* updates library version from 3.0 to 3.1
+* bug fix: year was not properly saved after retrieving metadata
+* add menu for relaxed or compact view
+
 ## v1.3.2
 * fix bug preventing moving and deleting of multiple entries